remove almost every device from the device tree
Wipe all the compatible strings except for the uarts, and also wipe the device_type and name properties
for a few devices that still matches kexts after the wipe.

This avoids loading most of the drivers, and gets boot to progress further.

This gets the kernel to "Still waiting for root device", but with a 30 second delay during boot since
it tries to wait for a nonexistant RTC device (pmu,d2422, which is removed from the device tree)